+++ CHANGELOG.md
@@ -67,6 +67,7 @@
 - fix: export types from index to simplify import declarations
 - fix: do not mark form as validated when `validateFields()` was successful
 - fix: fix lost of modified state after using `appendListItem()`
+- fix: fix "undefined path error" when field onBlur is called
 
 ## v5.6.0 (2025-01-09)